+++ ton_client/src/abi/encode_message.rs
@@ -10,7 +10,7 @@ use serde_json::Value;
 use std::str::FromStr;
 use std::sync::Arc;
 use ton_abi::Contract;
-use ton_block::{MsgAddressInt, CurrencyCollection};
+use ton_block::{MsgAddressInt, MsgAddressIntOrNone, CurrencyCollection};
 use ton_sdk::{ContractImage, FunctionCallSet};
 
 //--------------------------------------------------------------------------- encode_deploy_message
@@ -277,6 +277,7 @@ fn encode_deploy(
 }
 
 fn encode_int_deploy(
+    src: MsgAddressIntOrNone,
     context: std::sync::Arc<ClientContext>,
     abi: &str,
     image: ContractImage,
@@ -288,6 +289,7 @@ fn encode_int_deploy(
 ) -> ClientResult<(Vec<u8>, MsgAddressInt)> {
     let address = image.msg_address(workchain_id);
     let message = ton_sdk::Contract::get_int_deploy_message_bytes(
+        src,
         call_set.to_function_call_set(pubkey, None, &context, &abi, true)?,
         image,
         workchain_id,
@@ -317,6 +319,7 @@ fn encode_empty_deploy(
 }
 
 fn encode_empty_int_deploy(
+    src: MsgAddressIntOrNone,
     image: ContractImage,
     workchain_id: i32,
     ihr_disabled: bool,
@@ -324,6 +327,7 @@ fn encode_empty_int_deploy(
 ) -> ClientResult<(Vec<u8>, MsgAddressInt)> {
     let address = image.msg_address(workchain_id);
     let message = ton_sdk::Contract::construct_int_deploy_message_no_constructor(
+        src,
         image,
         workchain_id,
         ihr_disabled,
@@ -473,13 +477,16 @@ pub async fn encode_message(
 
 #[derive(Serialize, Deserialize, Clone, Debug, ApiType, Default)]
 pub struct ParamsOfEncodeInternalMessage {
-    /// Contract ABI.
-    pub abi: Abi,
+    /// Contract ABI. Can be None if both deploy_set and call_set are None.
+    pub abi: Option<Abi>,
 
     /// Target address the message will be sent to.
     ///
     /// Must be specified in case of non-deploy message.
     pub address: Option<String>,
+
+    /// Source address of the message.
+    pub src_address: Option<String>,
 
     /// Deploy parameters.
     ///
@@ -494,7 +501,7 @@ pub struct ParamsOfEncodeInternalMessage {
     /// of the functions that will to be called upon deploy transaction.
     pub call_set: Option<CallSet>,
 
-    /// Value in nanograms to be sent with message.
+    /// Value in nanotokens to be sent with message.
     pub value: String,
 
     /// Flag of bounceable message. Default is true.
@@ -538,12 +545,18 @@ pub async  fn encode_internal_message(
     context: std::sync::Arc<ClientContext>,
     params: ParamsOfEncodeInternalMessage,
 ) -> ClientResult<ResultOfEncodeInternalMessage> {
-    let abi = params.abi.json_string()?;
-
+    let src_address = match params.src_address {
+        Some(addr) => MsgAddressIntOrNone::Some(account_decode(&addr)?),
+        None => MsgAddressIntOrNone::None,
+    };
     let ihr_disabled = !params.enable_ihr.unwrap_or(false);
     let bounce = params.bounce.unwrap_or(true);
 
     let (message, address) = if let Some(deploy_set) = params.deploy_set {
+        let abi = params.abi
+            .ok_or_else(|| Error::invalid_abi("abi is undefined"))?
+            .json_string()?;
+
         let workchain_id = deploy_set
             .workchain_id
             .unwrap_or(context.config.abi.workchain);
@@ -558,6 +571,7 @@ pub async  fn encode_internal_message(
         let public = required_public_key(public)?;
         if let Some(call_set) = &params.call_set {
             encode_int_deploy(
+                src_address,
                 Arc::clone(&context),
                 &abi,
                 image,
@@ -568,30 +582,45 @@ pub async  fn encode_internal_message(
                 bounce,
             )?
         } else {
-            encode_empty_int_deploy(image, workchain_id, ihr_disabled, bounce)?
+            encode_empty_int_deploy(src_address, image, workchain_id, ihr_disabled, bounce)?
         }
-    } else if let Some(call_set) = &params.call_set {
+    } else {
         let address = params
             .address
             .as_ref()
             .ok_or(abi::Error::required_address_missing_for_encode_message())?;
         let address = account_decode(address)?;
+        let value = CurrencyCollection::with_grams(
+            u64::from_str(&params.value)
+                .map_err(|err| abi::Error::encode_run_message_failed(err, ""))?
+        );
+        if let Some(call_set) = &params.call_set {
+            let abi = params.abi
+                .ok_or_else(|| Error::invalid_abi("abi is undefined"))?
+                .json_string()?;
+            let message = ton_sdk::Contract::construct_call_int_message_json(
+                address.clone(),
+                src_address,
+                ihr_disabled,
+                bounce,
+                value,
+                call_set.to_function_call_set(None, None, &context, &abi, true)?,
+            )
+            .map_err(|err| abi::Error::encode_run_message_failed(err, &call_set.function_name))?;
 
-        let message = ton_sdk::Contract::construct_call_int_message_json(
-            address.clone(),
-            ihr_disabled,
-            bounce,
-            CurrencyCollection::with_grams(
-                u64::from_str(&params.value)
-                    .map_err(|err| abi::Error::encode_run_message_failed(err, ""))?
-            ),
-            call_set.to_function_call_set(None, None, &context, &abi, true)?,
-        )
-        .map_err(|err| abi::Error::encode_run_message_failed(err, &call_set.function_name))?;
-
-        (message.serialized_message, address)
-    } else {
-        return Err(abi::Error::missing_required_call_set_for_encode_message());
+            (message.serialized_message, address)
+        } else {
+            let message = ton_sdk::Contract::construct_int_message_with_body(
+                address.clone(),
+                src_address,
+                ihr_disabled,
+                bounce,
+                value,
+                None,
+            )
+            .map_err(|err| abi::Error::encode_run_message_failed(err, ""))?;
+            (message.serialized_message, address)
+        }
     };
 
     Ok(ResultOfEncodeInternalMessage {