+++ ton_client/src/json_interface/modules.rs
@@ -14,6 +14,7 @@
 
 use super::registrar::ModuleReg;
 use super::runtime::RuntimeHandlers;
+use crate::boc::BuilderOp;
 
 /// Provides information about library.
 #[derive(ApiModule)]
@@ -326,6 +327,11 @@ fn register_boc(handlers: &mut RuntimeHandlers) {
     module.register_async_fn(
         crate::boc::cache_unpin,
         crate::boc::cache::cache_unpin_api,
+    );
+    module.register_type::<BuilderOp>();
+    module.register_async_fn(
+        crate::boc::encode_boc,
+        crate::boc::encode::encode_boc_api,
     );
     module.register();
 }